+++ snakemake/__init__.py
@@ -6,7 +6,7 @@ __license__ = "MIT"
 import os
 import subprocess
 import glob
-from argparse import ArgumentError
+from argparse import ArgumentError, ArgumentDefaultsHelpFormatter
 import logging as _logging
 import re
 import sys
@@ -42,6 +42,7 @@ def snakemake(
     batch=None,
     cache=None,
     report=None,
+    report_stylesheet=None,
     lint=None,
     listrules=False,
     list_target_rules=False,
@@ -88,7 +89,7 @@ def snakemake(
     lock=True,
     unlock=False,
     cleanup_metadata=None,
-    cleanup_conda=False,
+    conda_cleanup_envs=False,
     cleanup_shadow=False,
     cleanup_scripts=True,
     force_incomplete=False,
@@ -131,10 +132,11 @@ def snakemake(
     use_env_modules=False,
     singularity_args="",
     conda_prefix=None,
+    conda_cleanup_pkgs=None,
     list_conda_envs=False,
     singularity_prefix=None,
     shadow_prefix=None,
-    create_envs_only=False,
+    conda_create_envs_only=False,
     mode=Mode.default,
     wrapper_prefix=None,
     kubernetes=None,
@@ -151,6 +153,7 @@ def snakemake(
     show_failed_logs=False,
     keep_incomplete=False,
     messaging=None,
+    edit_notebook=None,
 ):
     """Run snakemake on a given snakefile.
 
@@ -200,7 +203,7 @@ def snakemake(
         lock (bool):                lock the working directory when executing the workflow (default True)
         unlock (bool):              just unlock the working directory (default False)
         cleanup_metadata (list):    just cleanup metadata of given list of output files (default None)
-        cleanup_conda (bool):       just cleanup unused conda environments (default False)
+        conda_cleanup_envs (bool):  just cleanup unused conda environments (default False)
         cleanup_shadow (bool):      just cleanup old shadow directories (default False)
         cleanup_scripts (bool):     delete wrapper scripts used for execution (default True)
         force_incomplete (bool):    force the re-creation of incomplete files (default False)
@@ -240,9 +243,11 @@ def snakemake(
         use_env_modules (bool):     load environment modules if defined in rules
         singularity_args (str):     additional arguments to pass to singularity
         conda_prefix (str):         the directory in which conda environments will be created (default None)
+        conda_cleanup_pkgs (snakemake.deployment.conda.CondaCleanupMode):
+                                    whether to clean up conda tarballs after env creation (default None), valid values: "tarballs", "cache"
         singularity_prefix (str):   the directory to which singularity images will be pulled (default None)
         shadow_prefix (str):        prefix for shadow directories. The job-specific shadow directories will be created in $SHADOW_PREFIX/shadow/ (default None)
-        create_envs_only (bool):    if specified, only builds the conda environments specified for each job, then exits.
+        conda_create_envs_only (bool):    if specified, only builds the conda environments specified for each job, then exits.
         list_conda_envs (bool):     list conda environments and their location on disk.
         mode (snakemake.common.Mode): execution mode
         wrapper_prefix (str):       prefix for wrapper script URLs (default None)
@@ -258,7 +263,8 @@ def snakemake(
         cluster_status (str):       status command for cluster execution. If None, Snakemake will rely on flag files. Otherwise, it expects the command to return "success", "failure" or "running" when executing with a cluster jobid as single argument.
         export_cwl (str):           Compile workflow to CWL and save to given file
         log_handler (function):     redirect snakemake output to this custom log handler, a function that takes a log message dictionary (see below) as its only argument (default None). The log message dictionary for the log handler has to following entries:
-        keep_incomplete (bool):      keep incomplete output files of failed jobs
+        keep_incomplete (bool):     keep incomplete output files of failed jobs
+        edit_notebook (object):     "notebook.Listen" object to configuring notebook server for interactive editing of a rule notebook. If None, do not edit.
         log_handler (list):         redirect snakemake output to this list of custom log handler, each a function that takes a log message dictionary (see below) as its only argument (default []). The log message dictionary for the log handler has to following entries:
 
             :level:
@@ -330,10 +336,10 @@ def snakemake(
     if updated_files is None:
         updated_files = list()
 
-    if cluster or cluster_sync or drmaa or tibanna:
-        cores = sys.maxsize
+    if cluster or cluster_sync or drmaa or tibanna or kubernetes:
+        cores = None
     else:
-        nodes = sys.maxsize
+        nodes = None
 
     if isinstance(cluster_config, str):
         # Loading configuration from one file is still supported for
@@ -351,9 +357,15 @@ def snakemake(
         cluster_config_content = dict()
 
     run_local = not (cluster or cluster_sync or drmaa or kubernetes or tibanna)
-    if run_local and not dryrun:
-        # clean up all previously recorded jobids.
-        shell.cleanup()
+    if run_local:
+        if not dryrun:
+            # clean up all previously recorded jobids.
+            shell.cleanup()
+    else:
+        if edit_notebook:
+            raise WorkflowError(
+                "Notebook edit mode is only allowed with local execution."
+            )
 
     # force thread use for any kind of cluster
     use_threads = (
@@ -466,10 +478,11 @@ def snakemake(
             config_args=config_args,
             debug=debug,
             verbose=verbose,
-            use_conda=use_conda or list_conda_envs or cleanup_conda,
+            use_conda=use_conda or list_conda_envs or conda_cleanup_envs,
             use_singularity=use_singularity,
             use_env_modules=use_env_modules,
             conda_prefix=conda_prefix,
+            conda_cleanup_pkgs=conda_cleanup_pkgs,
             singularity_prefix=singularity_prefix,
             shadow_prefix=shadow_prefix,
             singularity_args=singularity_args,
@@ -486,6 +499,7 @@ def snakemake(
             cores=cores,
             nodes=nodes,
             resources=resources,
+            edit_notebook=edit_notebook,
         )
         success = True
         workflow.include(
@@ -495,7 +509,7 @@ def snakemake(
 
         if not print_compilation:
             if lint:
-                workflow.lint(json=lint == "json")
+                success = not workflow.lint(json=lint == "json")
             elif listrules:
                 workflow.list_rules()
             elif list_target_rules:
@@ -531,7 +545,7 @@ def snakemake(
                     lock=lock,
                     unlock=unlock,
                     cleanup_metadata=cleanup_metadata,
-                    cleanup_conda=cleanup_conda,
+                    conda_cleanup_envs=conda_cleanup_envs,
                     cleanup_shadow=cleanup_shadow,
                     cleanup_scripts=cleanup_scripts,
                     force_incomplete=force_incomplete,
@@ -554,13 +568,14 @@ def snakemake(
                     use_singularity=use_singularity,
                     use_env_modules=use_env_modules,
                     conda_prefix=conda_prefix,
+                    conda_cleanup_pkgs=conda_cleanup_pkgs,
                     singularity_prefix=singularity_prefix,
                     shadow_prefix=shadow_prefix,
                     singularity_args=singularity_args,
                     list_conda_envs=list_conda_envs,
                     kubernetes=kubernetes,
                     container_image=container_image,
-                    create_envs_only=create_envs_only,
+                    conda_create_envs_only=conda_create_envs_only,
                     default_remote_provider=default_remote_provider,
                     default_remote_prefix=default_remote_prefix,
                     tibanna=tibanna,
@@ -629,7 +644,7 @@ def snakemake(
                     nodeps=nodeps,
                     keep_target_files=keep_target_files,
                     cleanup_metadata=cleanup_metadata,
-                    cleanup_conda=cleanup_conda,
+                    conda_cleanup_envs=conda_cleanup_envs,
                     cleanup_shadow=cleanup_shadow,
                     cleanup_scripts=cleanup_scripts,
                     subsnakemake=subsnakemake,
@@ -638,10 +653,11 @@ def snakemake(
                     greediness=greediness,
                     no_hooks=no_hooks,
                     force_use_threads=use_threads,
-                    create_envs_only=create_envs_only,
+                    conda_create_envs_only=conda_create_envs_only,
                     assume_shared_fs=assume_shared_fs,
                     cluster_status=cluster_status,
                     report=report,
+                    report_stylesheet=report_stylesheet,
                     export_cwl=export_cwl,
                     batch=batch,
                     keepincomplete=keep_incomplete,
@@ -807,6 +823,7 @@ def get_argument_parser(profile=None):
     parser = configargparse.ArgumentParser(
         description="Snakemake is a Python based language and execution "
         "environment for GNU Make-like workflows.",
+        formatter_class=ArgumentDefaultsHelpFormatter,
         default_config_files=config_files,
         config_file_parser_class=YAMLConfigFileParser,
     )
@@ -851,7 +868,7 @@ def get_argument_parser(profile=None):
 
     group_exec.add_argument(
         "--cache",
-        nargs="+",
+        nargs="*",
         metavar="RULE",
         help="Store output files of given rules in a central cache given by the environment "
         "variable $SNAKEMAKE_OUTPUT_CACHE. Likewise, retrieve output files of the given rules "
@@ -908,10 +925,10 @@ def get_argument_parser(profile=None):
         help=(
             "Define additional resources that shall constrain the scheduling "
             "analogously to threads (see above). A resource is defined as "
-            "a name and an integer value. E.g. --resources gpu=1. Rules can "
+            "a name and an integer value. E.g. --resources mem_mb=1000. Rules can "
             "use resources by defining the resource keyword, e.g. "
-            "resources: gpu=1. If now two rules require 1 of the resource "
-            "'gpu' they won't be run in parallel by the scheduler."
+            "resources: mem_mb=600. If now two rules require 600 of the resource "
+            "'mem_mb' they won't be run in parallel by the scheduler."
         ),
     )
     group_exec.add_argument(
@@ -1085,16 +1102,46 @@ def get_argument_parser(profile=None):
         ),
     )
 
-    group_utils = parser.add_argument_group("UTILITIES")
+    group_report = parser.add_argument_group("REPORTS")
 
-    group_utils.add_argument(
+    group_report.add_argument(
         "--report",
         nargs="?",
         const="report.html",
-        metavar="HTMLFILE",
+        metavar="FILE",
         help="Create an HTML report with results and statistics. "
-        "If no filename is given, report.html is the default.",
+        "This can be either a .html file or a .zip file. "
+        "In the former case, all results are embedded into the .html (this only works for small data). "
+        "In the latter case, results are stored along with a file report.html in the zip archive. "
+        "If no filename is given, an embedded report.html is the default.",
     )
+    group_report.add_argument(
+        "--report-stylesheet",
+        metavar="CSSFILE",
+        help="Custom stylesheet to use for report. In particular, this can be used for "
+        "branding the report with e.g. a custom logo, see docs.",
+    )
+
+    group_notebooks = parser.add_argument_group("NOTEBOOKS")
+
+    group_notebooks.add_argument(
+        "--edit-notebook",
+        metavar="TARGET",
+        help="Interactively edit the notebook associated with the rule used to generate the given target file. "
+        "This will start a local jupyter notebook server. "
+        "Any changes to the notebook should be saved, and the server has to be stopped by "
+        "closing the notebook and hitting the 'Quit' button on the jupyter dashboard. "
+        "Afterwards, the updated notebook will be automatically stored in the path defined in the rule. "
+        "If the notebook is not yet present, this will create an empty draft. ",
+    )
+    group_notebooks.add_argument(
+        "--notebook-listen",
+        metavar="IP:PORT",
+        default="localhost:8888",
+        help="The IP address and PORT the notebook server used for editing the notebook (--edit-notebook) will listen on.",
+    )
+
+    group_utils = parser.add_argument_group("UTILITIES")
     group_utils.add_argument(
         "--lint",
         nargs="?",
@@ -1104,6 +1151,7 @@ def get_argument_parser(profile=None):
         "specific suggestions to improve code quality (work in progress, more lints "
         "to be added in the future). If no argument is provided, plain text output is used.",
     )
+
     group_utils.add_argument(
         "--export-cwl",
         action="store",
@@ -1127,7 +1175,9 @@ def get_argument_parser(profile=None):
         action="store_true",
         help="Do not execute anything and print the directed "
         "acyclic graph of jobs in the dot language. Recommended "
-        "use on Unix systems: snakemake --dag | dot | display",
+        "use on Unix systems: snakemake --dag | dot | display"
+        "Note print statements in your Snakefile may interfere "
+        "with visualization.",
     )
     group_utils.add_argument(
         "--rulegraph",
@@ -1138,7 +1188,9 @@ def get_argument_parser(profile=None):
         "Note that each rule is displayed once, hence the displayed graph will be "
         "cyclic if a rule appears in several steps of the workflow. "
         "Use this if above option leads to a DAG that is too large. "
-        "Recommended use on Unix systems: snakemake --rulegraph | dot | display",
+        "Recommended use on Unix systems: snakemake --rulegraph | dot | display"
+        "Note print statements in your Snakefile may interfere "
+        "with visualization.",
     )
     group_utils.add_argument(
         "--filegraph",
@@ -1149,7 +1201,9 @@ def get_argument_parser(profile=None):
         "Note that each rule is displayed once, hence the displayed graph will be "
         "cyclic if a rule appears in several steps of the workflow. "
         "Use this if above option leads to a DAG that is too large. "
-        "Recommended use on Unix systems: snakemake --filegraph | dot | display",
+        "Recommended use on Unix systems: snakemake --filegraph | dot | display"
+        "Note print statements in your Snakefile may interfere "
+        "with visualization.",
     )
     group_utils.add_argument(
         "--d3dag",
@@ -1731,11 +1785,6 @@ def get_argument_parser(profile=None):
         help="List all conda environments and their location on " "disk.",
     )
     group_conda.add_argument(
-        "--cleanup-conda",
-        action="store_true",
-        help="Cleanup unused conda environments.",
-    )
-    group_conda.add_argument(
         "--conda-prefix",
         metavar="DIR",
         help="Specify a directory in which the 'conda' and 'conda-archive' "
@@ -1747,7 +1796,25 @@ def get_argument_parser(profile=None):
         "invocation directory, or as an absolute path.",
     )
     group_conda.add_argument(
-        "--create-envs-only",
+        "--conda-cleanup-envs",
+        action="store_true",
+        help="Cleanup unused conda environments.",
+    )
+    from snakemake.deployment.conda import CondaCleanupMode
+
+    group_conda.add_argument(
+        "--conda-cleanup-pkgs",
+        type=CondaCleanupMode,
+        const=CondaCleanupMode.tarballs,
+        choices=list(CondaCleanupMode),
+        nargs="?",
+        help="Cleanup conda packages after creating environments. "
+        "In case of 'tarballs' mode, will clean up all downloaded package tarballs. "
+        "In case of 'cache' mode, will additionally clean up unused package caches. "
+        "If mode is omitted, will default to only cleaning up the tarballs.",
+    )
+    group_conda.add_argument(
+        "--conda-create-envs-only",
         action="store_true",
         help="If specified, only creates the job-specific "
         "conda environments then exits. The `--use-conda` "
@@ -1819,6 +1886,8 @@ def main(argv=None):
             args.cluster_sync = adjust_path(args.cluster_sync)
         if args.cluster_status:
             args.cluster_status = adjust_path(args.cluster_status)
+        if args.report_stylesheet:
+            args.report_stylesheet = adjust_path(args.report_stylesheet)
 
     if args.bash_completion:
         cmd = b"complete -o bashdefault -C snakemake-bash-completion snakemake"
@@ -1871,6 +1940,7 @@ def main(argv=None):
         or args.rulegraph
         or args.summary
         or args.lint
+        or args.report
     )
 
     if args.cores is not None:
@@ -1924,7 +1994,7 @@ def main(argv=None):
         )
         sys.exit(1)
 
-    if (args.conda_prefix or args.create_envs_only) and not args.use_conda:
+    if (args.conda_prefix or args.conda_create_envs_only) and not args.use_conda:
         print(
             "Error: --use-conda must be set if --conda-prefix or "
             "--create-envs-only is set.",
@@ -2063,11 +2133,19 @@ def main(argv=None):
             slack_logger = logging.SlackLogger()
             log_handler.append(slack_logger.log_handler)
 
+        if args.edit_notebook:
+            from snakemake import notebook
+
+            args.target = [args.edit_notebook]
+            args.force = True
+            args.edit_notebook = notebook.Listen(args.notebook_listen)
+
         success = snakemake(
             args.snakefile,
             batch=batch,
             cache=args.cache,
             report=args.report,
+            report_stylesheet=args.report_stylesheet,
             lint=args.lint,
             listrules=args.list,
             list_target_rules=args.list_target_rules,
@@ -2119,7 +2197,7 @@ def main(argv=None):
             lock=not args.nolock,
             unlock=args.unlock,
             cleanup_metadata=args.cleanup_metadata,
-            cleanup_conda=args.cleanup_conda,
+            conda_cleanup_envs=args.conda_cleanup_envs,
             cleanup_shadow=args.cleanup_shadow,
             cleanup_scripts=not args.skip_script_cleanup,
             force_incomplete=args.rerun_incomplete,
@@ -2154,13 +2232,14 @@ def main(argv=None):
             force_use_threads=args.force_use_threads,
             use_conda=args.use_conda,
             conda_prefix=args.conda_prefix,
+            conda_cleanup_pkgs=args.conda_cleanup_pkgs,
             list_conda_envs=args.list_conda_envs,
             use_singularity=args.use_singularity,
             use_env_modules=args.use_envmodules,
             singularity_prefix=args.singularity_prefix,
             shadow_prefix=args.shadow_prefix,
             singularity_args=args.singularity_args,
-            create_envs_only=args.create_envs_only,
+            conda_create_envs_only=args.conda_create_envs_only,
             mode=args.mode,
             wrapper_prefix=args.wrapper_prefix,
             default_remote_provider=args.default_remote_provider,
@@ -2170,6 +2249,7 @@ def main(argv=None):
             export_cwl=args.export_cwl,
             show_failed_logs=args.show_failed_logs,
             keep_incomplete=args.keep_incomplete,
+            edit_notebook=args.edit_notebook,
             log_handler=log_handler,
         )