+++ tests/tests.py
@@ -120,7 +120,16 @@ def test_ancient():
 
 
 def test_report():
-    run(dpath("test_report"), report="report.html", check_md5=False)
+    run(
+        dpath("test_report"),
+        report="report.html",
+        report_stylesheet="custom-stylesheet.css",
+        check_md5=False,
+    )
+
+
+def test_report_zip():
+    run(dpath("test_report_zip"), report="report.zip", check_md5=False)
 
 
 def test_dynamic():
@@ -237,6 +246,7 @@ def test_remote():
 
 
 def test_cluster_sync():
+    os.environ["TESTVAR"] = "test"
     run(dpath("test14"), snakefile="Snakefile.nonstandard", cluster_sync="./qsub")
 
 
@@ -960,7 +970,7 @@ def test_github_issue105():
 def test_output_file_cache():
     test_path = dpath("test_output_file_cache")
     os.environ["SNAKEMAKE_OUTPUT_CACHE"] = os.path.join(test_path, "cache")
-    run(test_path, cache=["a", "b", "c"])
+    run(test_path, cache=["a", "b"])
     run(test_path, cache=["invalid_multi"], targets="invalid1.txt", shouldfail=True)
 
 
@@ -1002,3 +1012,17 @@ def test_linting():
     snakemake(
         snakefile=os.path.join(dpath("test14"), "Snakefile.nonstandard"), lint=True
     )
+
+
+def test_string_resources():
+    from snakemake.resources import DefaultResources
+
+    run(
+        dpath("test_string_resources"),
+        default_resources=DefaultResources(["gpu_model='nvidia-tesla-1000'"]),
+        cluster="./qsub.py",
+    )
+
+
+def test_jupyter_notebook():
+    run(dpath("test_jupyter_notebook"), use_conda=True)