+++ CHANGELOG.rst
@@ -1,3 +1,62 @@
+[5.15.0] - 2020-04-21
+=====================
+Changed
+-------
+- The resource directive can now take strings, e.g. for defining a GPU model (see docs). This will e.g. be used for upcoming updates to cloud executors.
+- More extensive conda cleanup with --conda-cleanup-packages, meant for CI usage.
+- Further polish for reports.
+
+[5.14.0] - 2020-04-08
+=====================
+Changed
+-------
+- Redesigned HTML reports, with improved interface and performance.
+- For big data, HTML reports can now be stored as ZIP, where files are not anymore embedded but rather are stored in an auxilliary folder, such that they don't have to be in memory during report rendering.
+- Added subcategories to report (see docs).
+- Fixed a bug linter, leading to only one rule or snakefile to be linted.
+- Breaking change in CLI: added flags --conda-cleanup-envs and --conda-cleanup-pkgs, removed flag --cleanup-conda.
+- Fixed scheduling of pipe jobs, they are now always scheduled, fixing a hangup.
+- Corrected quoting of shell command for cluster submission.
+
+[5.13.0] - 2020-03-27
+=====================
+Added
+-----
+- Allow to flag directories for inclusion in the report.
+Changed
+-------
+- Fixed hash computation for --cache in case of positional params arguments.
+- Automatically restrict thread usage of linear algebra libraries to whatever is specified in the rule/job.
+
+[5.12.3] - 2020-03-24
+=====================
+Changed
+-------
+- Various minor bug fixes.
+
+[5.12.2] - 2020-03-24
+=====================
+Changed
+-------
+- Further improved linter output.
+
+[5.12.1] - 2020-03-24
+=====================
+Changed
+-------
+- Linter fixes
+
+[5.12.0] - 2020-03-24
+=====================
+Changed
+-------
+- Fixed the ability to supply functions for the thread directive.
+- Improved error messages for caching.
+
+Added
+-----
+- A new "cache: true" directive that allows to annotate between workflow caching eligibility for rules in the workflow.
+
 [5.11.2] - 2020-03-19
 =====================
 Changed